BASIC_PATH_TRACER: Explicitly copy the camera to device memory
CUDA does not like structs over a certain size (128bytes?) to be passed in
as args to the kernel. However, it doesn't give any warnings about it, but rather,
anything over this size is filled with bogus data.
To fix the problem, we allocate space in global memory for it, and copy over the
data from the CPU whenever the camera moves.